added if statement for checking if compile is needed

+++ linux_startup.sh
@@ -1,6 +1,10 @@
 #if you run this hacky script in rc.local for example the sandbox will start at boot.
-git fetch --all
-git reset --hard origin/master
-make
+if [ $(/usr/local/bin/git status) == *"Your branch is up-to-date with"* ] 
+then
+    git fetch --all
+    git reset --hard origin/master
+    make
+fi
+
 xinit &
 ./bin/sandboxSwarmDebug
